requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.10.1